Gyler.Domain.RPM.EVR: Introduce `Version` type
`Version` is a newtype wrapper around `NonEmptyText` with a smart constructor.
It can only contain Latin characters and certain separators:
> ['_', '.', '+', '~']

These characters are taken from the current version of `rpm-build` in ALT
(see the usage of `rpmCharCheck` in `build/parsePreamble.c`).

Upstream RPM also supports the '^' symbol.

`Version` is an instance of both `Serialize` and `Hashable`,
allowing it to be used with `unordered-containers` and `Gyler.FetchSpec`.

Additionally, `Version` has an instance of `IsText`, making it easy to
unwrap its value.

It is not entirely clear whether a `Version -> NonEmptyText` unwrapper
is necessary, so for now, `Version` does not supports it.