Gyler.Domain.RPM: Introduce the RPM type
Currently, it contains only the Name and EVR fields. In the future, this
module might be renamed to RPM.NEVR if the RPM representation requires
additional metadata.

The `_evr` field has the type `Maybe EVR` because there are cases where the
only known information about an RPM package is its name. For example, in
the Girar `del` subtask, an RPM package is represented as `del=PKGNAME`.

No parser is provided for the RPM type. Instead, it is recommended to use
`nameP` and `evrP` from `RPM.Name` and `RPM.EVR`.