Gyler.Domain.RPM.*: Add Renderable instances
Most of them are simple newtype derivations.

The only slightly more complex Renderable instance is implemented for
the EVR type. It hides the `Epoch` field when its value is zero. The
`showEVR` function, which had the same behavior and was previously
used in tests, has been replaced with `render`.

Currently, the RPM type does not have a `Renderable` instance.
There are several possible ways to implement it, but none of them seem
strictly necessary at the moment.